# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"AuthorDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"profilePicture": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CommentDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"cardId": {
						"type": "string"
					},
					"author": {
						"$ref": "#/components/schemas/AuthorDto"
					},
					"content": {
						"type": "string"
					},
					"likes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"cardId",
					"author",
					"content",
					"likes",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCommentDto": {
				"properties": {
					"cardId": {
						"type": "string"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"cardId",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CardVisibilityEnum": {
				"enum": [
					"private",
					"public",
					"unlisted"
				],
				"type": "string"
			},
			"CardDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"mediaUrls": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"visibility": {
						"$ref": "#/components/schemas/CardVisibilityEnum"
					},
					"author": {
						"$ref": "#/components/schemas/AuthorDto"
					},
					"isLiked": {
						"type": "boolean"
					},
					"isFavorited": {
						"type": "boolean"
					},
					"likes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"favorites": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"comments": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"category": {
						"type": "string"
					},
					"gradient": {
						"type": "string"
					},
					"allowComments": {
						"type": "boolean"
					},
					"allowDownloads": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"mediaUrls",
					"visibility",
					"author",
					"likes",
					"favorites",
					"comments",
					"tags",
					"category",
					"gradient",
					"allowComments",
					"allowDownloads",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResponseDto_CardDto_": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/CardDto"
						},
						"type": "array"
					},
					"nextCursor": {
						"type": "string"
					}
				},
				"required": [
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PopulatedCardDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"mediaUrls": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"visibility": {
						"$ref": "#/components/schemas/CardVisibilityEnum"
					},
					"author": {
						"$ref": "#/components/schemas/AuthorDto"
					},
					"isLiked": {
						"type": "boolean"
					},
					"isFavorited": {
						"type": "boolean"
					},
					"likes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"favorites": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"comments": {
						"items": {
							"$ref": "#/components/schemas/CommentDto"
						},
						"type": "array"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"category": {
						"type": "string"
					},
					"gradient": {
						"type": "string"
					},
					"allowComments": {
						"type": "boolean"
					},
					"allowDownloads": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"mediaUrls",
					"visibility",
					"author",
					"likes",
					"favorites",
					"comments",
					"tags",
					"category",
					"gradient",
					"allowComments",
					"allowDownloads",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateCardDto": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"mediaUrls": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"visibility": {
						"$ref": "#/components/schemas/CardVisibilityEnum"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"category": {
						"type": "string"
					},
					"gradient": {
						"type": "string"
					},
					"allowComments": {
						"type": "boolean"
					},
					"allowDownloads": {
						"type": "boolean"
					}
				},
				"required": [
					"title"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCardDto": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"mediaUrls": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"visibility": {
						"$ref": "#/components/schemas/CardVisibilityEnum"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"category": {
						"type": "string"
					},
					"gradient": {
						"type": "string"
					},
					"allowComments": {
						"type": "boolean"
					},
					"allowDownloads": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"GeneratedCardResponse": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"gradient": {
						"type": "string"
					},
					"visibility": {
						"type": "string",
						"enum": [
							"public"
						],
						"nullable": false
					},
					"allowComments": {
						"type": "boolean"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"description",
					"category",
					"gradient",
					"visibility",
					"allowComments",
					"tags"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RelatedCardDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"author": {
						"properties": {
							"profilePicture": {
								"type": "string"
							},
							"username": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"id"
						],
						"type": "object"
					},
					"gradient": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"title",
					"author",
					"gradient"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResponseDto_RelatedCardDto_": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/RelatedCardDto"
						},
						"type": "array"
					},
					"nextCursor": {
						"type": "string"
					}
				},
				"required": [
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"NotificationType": {
				"enum": [
					"card_like",
					"comment_like",
					"comment",
					"follow",
					"other"
				],
				"type": "string"
			},
			"NotificationDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/NotificationType"
					},
					"message": {
						"type": "string"
					},
					"sender": {
						"type": "string"
					},
					"recipient": {
						"type": "string"
					},
					"cardId": {
						"type": "string"
					},
					"read": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"type",
					"message",
					"sender",
					"recipient",
					"read",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "cardshare-api",
		"version": "1.0.0",
		"description": "CardShare API with Express.",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/comments": {
			"post": {
				"operationId": "CreateComment",
				"responses": {
					"201": {
						"description": "Comment created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentDto"
								}
							}
						}
					}
				},
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCommentDto"
							}
						}
					}
				}
			}
		},
		"/comments/{id}/like": {
			"post": {
				"operationId": "ToggleLikeComment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommentDto"
								}
							}
						}
					},
					"404": {
						"description": "Comment not found"
					}
				},
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/comments/{id}": {
			"delete": {
				"operationId": "DeleteComment",
				"responses": {
					"204": {
						"description": "Comment deleted"
					},
					"404": {
						"description": "Comment not found"
					}
				},
				"tags": [
					"Comments"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/cards": {
			"get": {
				"operationId": "GetCards",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponseDto_CardDto_"
								}
							}
						}
					}
				},
				"tags": [
					"Cards"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"latest",
								"most-liked"
							]
						}
					}
				]
			},
			"post": {
				"operationId": "CreateCard",
				"responses": {
					"201": {
						"description": "Card created",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/CardDto"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Cards"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCardDto"
							}
						}
					}
				}
			}
		},
		"/cards/{id}": {
			"get": {
				"operationId": "GetCardById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/PopulatedCardDto"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Card not found"
					}
				},
				"tags": [
					"Cards"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/CardDto"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Card not found"
					}
				},
				"tags": [
					"Cards"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCardDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCard",
				"responses": {
					"204": {
						"description": "Card deleted"
					},
					"404": {
						"description": "Card not found"
					}
				},
				"tags": [
					"Cards"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/cards/{id}/like": {
			"post": {
				"operationId": "ToggleLikeCard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/CardDto"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Cards"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/cards/{id}/favorite": {
			"post": {
				"operationId": "ToggleFavoriteCard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/CardDto"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Cards"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/cards/generate": {
			"post": {
				"operationId": "GenerateCard",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GeneratedCardResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Cards"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/cards/{id}/related": {
			"get": {
				"operationId": "GetRelatedCards",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponseDto_RelatedCardDto_"
								}
							}
						}
					},
					"404": {
						"description": "Card not found"
					}
				},
				"tags": [
					"Cards"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 3,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/notifications": {
			"get": {
				"operationId": "GetNotifications",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/NotificationDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/notifications/{id}/read": {
			"put": {
				"operationId": "MarkAsRead",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NotificationDto"
								}
							}
						}
					},
					"404": {
						"description": "Notification not found"
					}
				},
				"tags": [
					"Notifications"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/users/me/cards": {
			"get": {
				"operationId": "GetMyCards",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedResponseDto_CardDto_"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"latest",
								"most-liked"
							]
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}